Updated DAOs

+++ src/main/java/io/xico26/spotifum_neo/dao/UserDAOImpl.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 @Repository
 public class UserDAOImpl implements UserDAO {
-    private EntityManager em;
+    private final EntityManager em;
 
     @Autowired
     public UserDAOImpl(EntityManager em) {
@@ -18,76 +18,52 @@ public class UserDAOImpl implements UserDAO {
 
     @Override
     public User findById(int userId) {
-        EntityManager em = emf.createEntityManager();
-        User u = em.find(User.class, userId);
-        em.close();
-        return u;
+        return em.find(User.class, userId);
     }
 
     @Override
     public List<User> findAll() {
-        EntityManager em = emf.createEntityManager();
-        List<User> users = em.createQuery("FROM User", User.class).getResultList();
-        em.close();
-        return users;
+        return em.createQuery("FROM User", User.class).getResultList();
     }
 
     @Override
     public User findByUsername(String username) {
-        EntityManager em = emf.createEntityManager();
         try {
             TypedQuery<User> query = em.createQuery("SELECT u FROM User u WHERE u.username=:username", User.class);
             query.setParameter("username", username);
             return query.getSingleResult();
         } catch (NoResultException e) {
             return null;
-        } finally {
-            em.close();
         }
     }
 
     @Override
     public User findByEmail(String email) {
-        EntityManager em = emf.createEntityManager();
         try {
             TypedQuery<User> query = em.createQuery("SELECT u FROM User u WHERE u.email=:email", User.class);
             query.setParameter("email", email);
             return query.getSingleResult();
         } catch (NoResultException e) {
             return null;
-        } finally {
-            em.close();
         }
     }
 
     @Override
     public void save(User u) {
-        EntityManager em = emf.createEntityManager();
-        EntityTransaction tx = em.getTransaction();
-        tx.begin();
         em.persist(u);
-        tx.commit();
-        em.close();
     }
 
     @Override
-    public void delete(User u) {
-        EntityManager em = emf.createEntityManager();
-        EntityTransaction tx = em.getTransaction();
-        tx.begin();
-        em.remove(u);
-        tx.commit();
-        em.close();
+    public void deleteById(int id) {
+        User u = em.find(User.class, id);
+        if  (u != null) {
+            em.remove(u);
+        }
     }
 
     @Override
     public void update(User u) {
-        EntityManager em = emf.createEntityManager();
-        EntityTransaction tx = em.getTransaction();
-        tx.begin();
         em.merge(u);
-        tx.commit();
-        em.close();
     }
 
 }
